[feat] 시각화 로직 파일 OrderDetail로 이사
